requirements.txt
django==4.2.6
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2==2.9.9
djangorestframework-simplejwt==5.3.0
django-rest-passwordreset==1.3.0
